+++ src/bot/commands/helpers.py
@@ -33,6 +33,46 @@ def get_hero_data(hero_id):
             return hero
 
 
+def get_hero_by_name(hero_name):
+    hero_name = hero_name.lower()
+
+    hero_data_file = (
+        JSON_CONSTANT_DATA_FILE_DIR + JSON_CONSTANT_DATA_FILE_MAPPING.HERO_DATA.value
+    )
+    hero_json = read_json_file(hero_data_file)
+    for hero in hero_json:
+        if hero["localized_name"].lower() == hero_name:
+            return hero
+
+    hero_alias_file = (
+        JSON_CONSTANT_DATA_FILE_DIR + JSON_CONSTANT_DATA_FILE_MAPPING.HERO_ALIASES.value
+    )
+    alias_json = read_json_file(hero_alias_file)
+    for hero in alias_json:
+        for alias in hero["aliases"]:
+            if alias.lower() == hero_name:
+                found_hero = get_hero_data(hero["id"])
+                return found_hero
+
+
+
+def filter_hero_winrates(hero_data, hero_id):
+    for hero in hero_data:
+        if hero["hero_id"] == hero_id:
+            return hero
+
+
+def format_winrate_response(hero_data, telegram_handle):
+    hero_by_id = get_hero_data(int(hero_data["hero_id"]))
+    hero_name = hero_by_id["localized_name"]
+    if hero_data["games"] == 0:
+        return f"@{telegram_handle} has no games as {hero_name} recorded"
+
+    winrate = hero_data["win"] / hero_data["games"] * 100
+    winrate = round(winrate, 3)
+    return f"@{telegram_handle} has a {winrate}% winrate as {hero_name} ({hero_data['win']} wins over {hero_data['games']} games)"
+
+
 def escape_markdown(text):
     return re.escape(text)